requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
python-dotenv==1.0.0
pyyaml==6.0.1
selenium==4.15.2
webdriver-manager==4.0.1
pytz==2023.3
pytest==7.4.3

